Link to CommonMark website in README.md
Link to the CommonMark website to explicitly reference what it is, and to help users unfamiliar with it to discover and navigate to it.

CommonMark is mentioned very early in the README. Here it is a subheadline. I decided to not link here but in this features section instead where it is listed as one of the features. The subheadline seems more fitting as a short, pure text for those knowing about it.

The link is to the main website rather than the specification as users unfamiliar with it will want to land on the landing page with a description of what it is rather than a technical specification. The specification page does not immediately and obviously link back to the main website.

The shortened term spec is being extended to its full word specification for clarity and discoverability.