renderer: Don't panic for unknown NodeKind
When rendering a Node, Renderer looks up the NodeRenderer for that
node's kind in a slice. That will panic with an "index out of range"
error if a node renderer for that node kind has not been registered.

This change verifies that a node renderer is available for that node
kind before attempting to use it, and instead of panicking, returns an
informative error message.