testutil: Add ParseTestCaseFile
This moves the logic for parsing a test case file
into a separate function named ParseTestCaseFile.
This will make it easier for extensions
to use the same format for test files,
even if they want a different strategy to run them.

The function does not do filtering like DoTestCaseFile;
that logic has been left inside DoTestCaseFile.

Minus that, this function does not modify any logic.
The next commits do.